requirements.txt
openai>=1.12.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0